+++ src/main/java/com/example/japapp/config/AppConfig.java
@@ -1,7 +1,7 @@
 package com.example.japapp.config;
 
-import com.example.japapp.filters.RoleFilter;
-import com.example.japapp.services.impl.UsersService;
+import com.example.japapp.filter.RoleFilter;
+import com.example.japapp.service.impl.UsersService;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
