+++ src/main/java/com/example/japapp/exception/MainException.java
@@ -1,0 +1,7 @@
+package com.example.japapp.exception;
+
+public class MainException extends RuntimeException {
+    public MainException(String message) {
+        super(message);
+    }
+}
