--- src/main/java/com/example/japapp/exceptions/MainException.java
@@ -1,7 +1,0 @@
-package com.example.japapp.exceptions;
-
-public class MainException extends RuntimeException {
-    public MainException(String message) {
-        super(message);
-    }
-}
