+++ src/main/java/com/example/japapp/model/Role.java
@@ -1,0 +1,60 @@
+package com.example.japapp.model;
+
+
+import jakarta.persistence.*;
+
+@Entity(name = "roles")
+@Table(
+        name = "roles"
+)
+public class Role {
+    @Id
+    @SequenceGenerator(
+            name = "role_sequence",
+            sequenceName = "role_sequence",
+            allocationSize = 1
+    )
+    @GeneratedValue(
+            strategy = GenerationType.SEQUENCE,
+            generator = "role_sequence"
+    )
+    @Column(
+            name = "role_id",
+            updatable = false
+    )
+    private Long id;
+
+    @Column(
+            name = "role_name",
+            nullable = false
+    )
+    private String name;
+
+    public Role(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public Role(String name) {
+        this.name = name;
+    }
+
+    public Role() {
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
