--- src/main/java/com/example/japapp/models/Book.java
@@ -1,71 +1,0 @@
-package com.example.japapp.models;
-
-import jakarta.persistence.*;
-
-@Entity(name = "books")
-@Table(
-        name = "books",
-        uniqueConstraints = {
-                @UniqueConstraint(name = "book_title_unique", columnNames = "title")
-        }
-)
-public class Book {
-    @Id
-    @SequenceGenerator(
-            name = "book_sequence",
-            sequenceName = "book_sequence",
-            allocationSize = 1
-    )
-    @GeneratedValue(
-            strategy = GenerationType.SEQUENCE,
-            generator = "book_sequence"
-    )
-    @Column(
-            name = "book_id",
-            updatable = false
-    )
-    private Long id;
-
-    @Column(
-            name = "title",
-            nullable = false
-    )
-    private String title;
-
-    @Column(
-            name = "annotation"
-    )
-    private String annotation;
-
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getAnnotation() {
-        return annotation;
-    }
-
-    public void setAnnotation(String annotation) {
-        this.annotation = annotation;
-    }
-}
