--- src/main/java/com/example/japapp/repositories/RolesRepository.java
@@ -1,10 +1,0 @@
-package com.example.japapp.repositories;
-
-import com.example.japapp.models.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RolesRepository extends JpaRepository<Role, Long> {
-    Optional<Role> findByName(String name);
-}
