--- src/main/java/com/example/japapp/repositories/UsersRepository.java
@@ -1,11 +1,0 @@
-package com.example.japapp.repositories;
-
-import com.example.japapp.models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UsersRepository extends JpaRepository<User, Long> {
-    User findByEmail(String email);
-    Optional<User> findById(long id);
-}
