+++ src/main/java/com/example/japapp/repository/RolesRepository.java
@@ -1,0 +1,10 @@
+package com.example.japapp.repository;
+
+import com.example.japapp.model.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RolesRepository extends JpaRepository<Role, Long> {
+    Optional<Role> findByName(String name);
+}
