+++ src/main/java/com/example/japapp/service/PasswordService.java
@@ -1,0 +1,6 @@
+package com.example.japapp.service;
+
+public interface PasswordService {
+    String hashPassword(String password);
+    boolean checkPassword(String password, String passwordHash);
+}
