--- src/main/java/com/example/japapp/services/PasswordService.java
@@ -1,6 +1,0 @@
-package com.example.japapp.services;
-
-public interface PasswordService {
-    String hashPassword(String password);
-    boolean checkPassword(String password, String passwordHash);
-}
