--- src/main/java/com/example/japapp/services/impl/AuthenticationService.java
@@ -1,25 +1,0 @@
-package com.example.japapp.services.impl;
-
-import com.example.japapp.models.User;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
-
-// TODO...
-
-@Service
-public class AuthenticationService {
-
-    public boolean isAuthenticated() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        return authentication != null && authentication.isAuthenticated();
-    }
-
-    public User getCurrentUser() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication == null || !authentication.isAuthenticated()) {
-            return null;
-        }
-        return (User) authentication.getPrincipal();
-    }
-}
