--- src/main/java/com/example/japapp/services/impl/PasswordServiceImpl.java
@@ -1,27 +1,0 @@
-package com.example.japapp.services.impl;
-
-import com.example.japapp.services.PasswordService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class PasswordServiceImpl implements PasswordService {
-
-    private final BCryptPasswordEncoder passwordEncoder;
-
-    @Autowired
-    public PasswordServiceImpl(BCryptPasswordEncoder passwordEncoder) {
-        this.passwordEncoder = passwordEncoder;
-    }
-
-    @Override
-    public String hashPassword(String password) {
-        return passwordEncoder.encode(password);
-    }
-
-    @Override
-    public boolean checkPassword(String password, String passwordHash) {
-        return passwordEncoder.matches(password, passwordHash);
-    }
-}
